executive-document-studio/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiosqlite==0.20.0
pydantic==2.9.2
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2
